Simplify ServerPoolClient
Yubico no longer uses an API which needs requests to be sent to
multiple endpoints, but rather has one geolocated API endpoint.
This removes the logic to try multiple and thereby simplifies
the code.

This is the most simplification we can do without breaking
the library's published interface. We can do that (would probably
mean to remove the SeverPoolClient altogether) but that breaks
implementers for no clear gain at this point. Just simplifying
the current code has a great benefit/cost tradeoff.